Structure change following @imrick proposal
