+++ App.tsx
@@ -1,75 +1,6 @@
 import React from "react";
-import { Ionicons } from "@expo/vector-icons";
-import { NavigationContainer } from "@react-navigation/native";
-import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
-import { createStackNavigator } from "@react-navigation/stack";
-import HomeScreen from "./navigation/HomeScreen";
-import DetailsScreen from "./navigation/DetailsScreen";
-import SettingsScreen from "./navigation/SettingsScreen";
-import { RootStackParamList } from "./navigation/Types";
-
-const HomeStack = createStackNavigator<RootStackParamList>();
-
-function HomeStackScreen() {
-  return (
-    <HomeStack.Navigator>
-      <HomeStack.Screen name="Home" component={HomeScreen} />
-      <HomeStack.Screen name="Details" component={DetailsScreen} />
-    </HomeStack.Navigator>
-  );
-}
-
-const SettingsStack = createStackNavigator<RootStackParamList>();
-
-function SettingsStackScreen() {
-  return (
-    <SettingsStack.Navigator>
-      <SettingsStack.Screen name="Settings" component={SettingsScreen} />
-      <SettingsStack.Screen name="Details" component={DetailsScreen} />
-    </SettingsStack.Navigator>
-  );
-}
-
-const Tab = createBottomTabNavigator<RootStackParamList>();
-
-function TabScreen() {
-  return (
-    <Tab.Navigator
-      screenOptions={({ route }) => ({
-        tabBarIcon: ({ focused, color, size }) => {
-          let iconName: any; // TODO: find better type
-
-          if (route.name === "Home") {
-            iconName = focused
-              ? "ios-information-circle"
-              : "ios-information-circle-outline";
-          } else if (route.name === "Settings") {
-            iconName = focused ? "ios-list-sharp" : "ios-list-outline";
-          }
-
-          // You can return any component that you like here!
-          return <Ionicons name={iconName} size={size} color={color} />;
-        },
-      })}
-      tabBarOptions={{
-        activeTintColor: "tomato",
-        inactiveTintColor: "gray",
-      }}
-    >
-      <Tab.Screen
-        name="Home"
-        component={HomeStackScreen}
-        options={{ tabBarBadge: 3 }}
-      />
-      <Tab.Screen name="Settings" component={SettingsStackScreen} />
-    </Tab.Navigator>
-  );
-}
+import { TabNavigator } from "./navigation/tab-navigation";
 
 export default function App() {
-  return (
-    <NavigationContainer>
-      <TabScreen></TabScreen>
-    </NavigationContainer>
-  );
+  return <TabNavigator></TabNavigator>;
 }
