+++ screens/DetailsScreen.tsx
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import { Text, View } from "react-native";
-import { DetailsScreenProps } from "./Types";
+import { DetailsScreenProps } from "../navigation/app-stacks";
 
 export default class DetailsScreen extends Component<DetailsScreenProps, {}> {
   render() {
